@@ -5,7 +5,7 @@ homepage: https://github.com/JackJonson/flutter_animated_dialog
 email: yswing1@gmail.com
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
