@@ -46,9 +46,9 @@ class _MyAppState extends State<MyApp> {
 class MyHomePage extends StatefulWidget {
   final String title;
 
-  final VoidCallback onSetting;
+  final VoidCallback? onSetting;
 
-  MyHomePage({Key key, this.title, this.onSetting}) : super(key: key);
+  MyHomePage({Key? key, required this.title, this.onSetting}) : super(key: key);
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -59,17 +59,17 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   // Whether the green box should be visible or invisible
 
-  String selectedIndexText;
+  String? selectedIndexText;
 
-  int selectIdx;
+  int? selectIdx;
 
-  String singleSelectedIndexText;
+  String? singleSelectedIndexText;
 
-  int selectIndex;
+  int? selectIndex;
 
-  String multiSelectedIndexesText;
+  String? multiSelectedIndexesText;
 
-  List<int> selectedIndexes;
+  List<int>? selectedIndexes;
 
   @override
   Widget build(BuildContext context) {
@@ -683,10 +683,10 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
             ListTile(
               title: Text(
-                "List dialog ${selectedIndexText != null && selectedIndexText.isNotEmpty ? '(index:' + selectedIndexText + ')' : ''}",
+                "List dialog ${selectedIndexText != null && selectedIndexText!.isNotEmpty ? '(index:' + selectedIndexText! + ')' : ''}",
               ),
               onTap: () async {
-                int index = await showAnimatedDialog(
+                int? index = await showAnimatedDialog(
                   context: context,
                   barrierDismissible: true,
                   builder: (BuildContext context) {
@@ -722,10 +722,10 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
             ListTile(
               title: Text(
-                "List single select${singleSelectedIndexText != null && singleSelectedIndexText.isNotEmpty ? '(index:' + singleSelectedIndexText + ')' : ''}",
+                "List single select${singleSelectedIndexText != null && singleSelectedIndexText!.isNotEmpty ? '(index:' + singleSelectedIndexText! + ')' : ''}",
               ),
               onTap: () async {
-                int index = await showAnimatedDialog(
+                int? index = await showAnimatedDialog(
                   context: context,
                   barrierDismissible: true,
                   builder: (BuildContext context) {
@@ -759,10 +759,10 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
             ListTile(
               title: Text(
-                "List multiple select${multiSelectedIndexesText != null && multiSelectedIndexesText.isNotEmpty ? '(index:' + multiSelectedIndexesText + ')' : ''}",
+                "List multiple select${multiSelectedIndexesText != null && multiSelectedIndexesText!.isNotEmpty ? '(index:' + multiSelectedIndexesText! + ')' : ''}",
               ),
               onTap: () async {
-                List<int> indexes = await showAnimatedDialog(
+                List<int>? indexes = await showAnimatedDialog(
                   context: context,
                   barrierDismissible: true,
                   builder: (BuildContext context) {
@@ -788,7 +788,7 @@ class _MyHomePageState extends State<MyHomePage> {
                 print('selectedIndex:${selectedIndexes?.toString()}');
                 setState(() {
                   this.multiSelectedIndexesText =
-                      selectedIndexes != null && selectedIndexes.length > 0
+                      selectedIndexes != null && selectedIndexes!.length > 0
                           ? selectedIndexes.toString()
                           : '';
                 });
