@@ -67,17 +67,16 @@ enum DialogTransitionType {
 }
 
 /// Displays a Material dialog above the current contents of the app
-Future<T> showAnimatedDialog<T>({
-  @required BuildContext context,
+Future<T?> showAnimatedDialog<T>({
+  required BuildContext context,
   bool barrierDismissible = false,
-  @required WidgetBuilder builder,
+  required WidgetBuilder builder,
   animationType = DialogTransitionType.fade,
   Curve curve = Curves.linear,
-  Duration duration,
-  AlignmentGeometry alignment = Alignment.center,
-  Axis axis,
+  Duration? duration,
+  Alignment alignment = Alignment.center,
+  Axis? axis,
 }) {
-  assert(builder != null);
   assert(debugCheckHasMaterialLocalizations(context));
 
   final ThemeData theme = Theme.of(context);
@@ -91,9 +90,7 @@ Future<T> showAnimatedDialog<T>({
       return SafeArea(
         top: false,
         child: Builder(builder: (BuildContext context) {
-          return theme != null
-              ? Theme(data: theme, child: pageChild)
-              : pageChild;
+          return Theme(data: theme, child: pageChild);
         }),
       );
     },
@@ -106,7 +103,6 @@ Future<T> showAnimatedDialog<T>({
       switch (animationType) {
         case DialogTransitionType.fade:
           return FadeTransition(opacity: animation, child: child);
-          break;
         case DialogTransitionType.slideFromRight:
           return SlideTransition(
             transformHitTests: false,
@@ -116,7 +112,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromLeft:
           return SlideTransition(
             transformHitTests: false,
@@ -126,7 +121,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromRightFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -138,7 +132,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.slideFromLeftFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -150,7 +143,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.slideFromTop:
           return SlideTransition(
             transformHitTests: false,
@@ -160,7 +152,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromTopFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -172,7 +163,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.slideFromBottom:
           return SlideTransition(
             transformHitTests: false,
@@ -182,7 +172,6 @@ Future<T> showAnimatedDialog<T>({
             ).chain(CurveTween(curve: curve)).animate(animation),
             child: child,
           );
-          break;
         case DialogTransitionType.slideFromBottomFade:
           return SlideTransition(
             position: Tween<Offset>(
@@ -194,7 +183,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.scale:
           return ScaleTransition(
             alignment: alignment,
@@ -208,7 +196,6 @@ Future<T> showAnimatedDialog<T>({
             ),
             child: child,
           );
-          break;
         case DialogTransitionType.fadeScale:
           return ScaleTransition(
             alignment: alignment,
@@ -228,7 +215,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.scaleRotate:
           return ScaleTransition(
             alignment: alignment,
@@ -247,7 +233,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.rotate:
           return CustomRotationTransition(
             alignment: alignment,
@@ -255,7 +240,6 @@ Future<T> showAnimatedDialog<T>({
                 parent: animation, curve: Interval(0.0, 1.0, curve: curve))),
             child: child,
           );
-          break;
         case DialogTransitionType.fadeRotate:
           return CustomRotationTransition(
             alignment: alignment,
@@ -269,7 +253,6 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.rotate3D:
           return Rotation3DTransition(
             alignment: alignment,
@@ -285,10 +268,9 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.size:
           return Align(
-            alignment: alignment ?? Alignment.center,
+            alignment: alignment,
             child: SizeTransition(
               sizeFactor: CurvedAnimation(
                 parent: animation,
@@ -298,10 +280,9 @@ Future<T> showAnimatedDialog<T>({
               child: child,
             ),
           );
-          break;
         case DialogTransitionType.sizeFade:
           return Align(
-            alignment: alignment ?? Alignment.center,
+            alignment: alignment,
             child: SizeTransition(
               sizeFactor: CurvedAnimation(
                 parent: animation,
@@ -316,10 +297,8 @@ Future<T> showAnimatedDialog<T>({
               ),
             ),
           );
-          break;
         case DialogTransitionType.none:
           return child;
-          break;
         default:
           return FadeTransition(opacity: animation, child: child);
       }
@@ -344,7 +323,7 @@ class CustomDialogWidget extends StatelessWidget {
   /// null, which implies a default that depends on the values of the other
   /// properties. See the documentation of [titlePadding] for details.
   const CustomDialogWidget({
-    Key key,
+    Key? key,
     this.title,
     this.titlePadding,
     this.titleTextStyle,
@@ -365,7 +344,7 @@ class CustomDialogWidget extends StatelessWidget {
   /// of the dialog.
   ///
   /// Typically a [Text] widget.
-  final Widget title;
+  final Widget? title;
 
   /// Padding around the title.
   ///
@@ -377,13 +356,13 @@ class CustomDialogWidget extends StatelessWidget {
   /// provided (but see [contentPadding]). If it _is_ null, then an extra 20
   /// pixels of bottom padding is added to separate the [title] from the
   /// [actions].
-  final EdgeInsetsGeometry titlePadding;
+  final EdgeInsetsGeometry? titlePadding;
 
   /// Style for the text in the [title] of this [AlertDialog].
   ///
   /// If null, [DialogTheme.titleTextStyle] is used, if that's null, defaults to
   /// [ThemeData.textTheme.title].
-  final TextStyle titleTextStyle;
+  final TextStyle? titleTextStyle;
 
   /// The (optional) content of the dialog is displayed in the center of the
   /// dialog in a lighter font.
@@ -392,7 +371,7 @@ class CustomDialogWidget extends StatelessWidget {
   /// message. As noted in the [AlertDialog] documentation, it's important
   /// to use a [SingleChildScrollView] if there's any risk that the content
   /// will not fit.
-  final Widget content;
+  final Widget? content;
 
   /// Padding around the content.
   ///
@@ -400,13 +379,13 @@ class CustomDialogWidget extends StatelessWidget {
   /// 20 pixels is provided above the content to separate the content from the
   /// title, and padding of 24 pixels is provided on the left, right, and bottom
   /// to separate the content from the other edges of the dialog.
-  final EdgeInsetsGeometry contentPadding;
+  final EdgeInsetsGeometry? contentPadding;
 
   /// Style for the text in the [content] of this [AlertDialog].
   ///
   /// If null, [DialogTheme.contentTextStyle] is used, if that's null, defaults
   /// to [ThemeData.textTheme.subhead].
-  final TextStyle contentTextStyle;
+  final TextStyle? contentTextStyle;
 
   /// The (optional) set of actions that are displayed at the bottom of the
   /// dialog.
@@ -419,17 +398,17 @@ class CustomDialogWidget extends StatelessWidget {
   /// If the [title] is not null but the [content] _is_ null, then an extra 20
   /// pixels of padding is added above the [ButtonBar] to separate the [title]
   /// from the [actions].
-  final List<Widget> actions;
+  final List<Widget>? actions;
 
   ///Widget in the bottom
-  final Widget bottomWidget;
+  final Widget? bottomWidget;
 
   /// {@macro flutter.material.dialog.backgroundColor}
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// {@macro flutter.material.dialog.elevation}
   /// {@macro flutter.material.material.elevation}
-  final double elevation;
+  final double? elevation;
 
   /// The semantic label of the dialog used by accessibility frameworks to
   /// announce screen transitions when the dialog is opened and closed.
@@ -442,13 +421,13 @@ class CustomDialogWidget extends StatelessWidget {
   ///
   ///  * [SemanticsConfiguration.isRouteName], for a description of how this
   ///    value is used.
-  final String semanticLabel;
+  final String? semanticLabel;
 
   /// {@macro flutter.material.dialog.shape}
-  final ShapeBorder shape;
+  final ShapeBorder? shape;
 
   ///Min width
-  final double minWidth;
+  final double? minWidth;
 
   @override
   Widget build(BuildContext context) {
@@ -456,7 +435,7 @@ class CustomDialogWidget extends StatelessWidget {
     final ThemeData theme = Theme.of(context);
     final DialogTheme dialogTheme = DialogTheme.of(context);
     final List<Widget> children = <Widget>[];
-    String label = semanticLabel;
+    var label = semanticLabel;
 
     if (title != null) {
       children.add(Padding(
@@ -465,7 +444,8 @@ class CustomDialogWidget extends StatelessWidget {
         child: DefaultTextStyle(
           style: titleTextStyle ??
               dialogTheme.titleTextStyle ??
-              theme.textTheme.headline6,
+              theme.textTheme.titleMedium ??
+              TextStyle(),
           child: Semantics(
             child: title,
             namesRoute: true,
@@ -481,18 +461,18 @@ class CustomDialogWidget extends StatelessWidget {
         case TargetPlatform.android:
         case TargetPlatform.fuchsia:
           label = semanticLabel ??
-              MaterialLocalizations.of(context)?.alertDialogLabel;
+              MaterialLocalizations.of(context).alertDialogLabel;
           break;
         case TargetPlatform.linux:
           label = semanticLabel ??
-              MaterialLocalizations.of(context)?.alertDialogLabel;
+              MaterialLocalizations.of(context).alertDialogLabel;
           break;
         case TargetPlatform.macOS:
           label = semanticLabel;
           break;
         case TargetPlatform.windows:
           label = semanticLabel ??
-              MaterialLocalizations.of(context)?.alertDialogLabel;
+              MaterialLocalizations.of(context).alertDialogLabel;
           break;
       }
     }
@@ -501,12 +481,13 @@ class CustomDialogWidget extends StatelessWidget {
       children.add(
         Flexible(
           child: Padding(
-            padding: contentPadding,
+            padding: contentPadding ?? EdgeInsets.zero,
             child: DefaultTextStyle(
               style: contentTextStyle ??
                   dialogTheme.contentTextStyle ??
-                  theme.textTheme.subtitle1,
-              child: content,
+                  theme.textTheme.titleSmall ??
+                  TextStyle(),
+              child: content!,
             ),
           ),
         ),
@@ -514,13 +495,13 @@ class CustomDialogWidget extends StatelessWidget {
     }
 
     if (bottomWidget != null) {
-      children.add(bottomWidget);
+      children.add(bottomWidget!);
     } else if (actions != null) {
       children.add(
         ButtonBarTheme(
           data: ButtonBarTheme.of(context),
           child: ButtonBar(
-            children: actions,
+            children: actions!,
           ),
         ),
       );
@@ -542,9 +523,11 @@ class CustomDialogWidget extends StatelessWidget {
       );
 
     dialogChild = CustomDialog(
-      backgroundColor: backgroundColor,
+      backgroundColor: backgroundColor ??
+          dialogTheme.backgroundColor ??
+          Theme.of(context).dialogBackgroundColor,
       elevation: elevation,
-      minWidth: minWidth,
+      minWidth: minWidth ?? 280.0,
       shape: shape,
       child: dialogChild,
     );
@@ -575,7 +558,7 @@ class CustomDialog extends StatelessWidget {
   ///
   /// Typically used in conjunction with [showDialog].
   const CustomDialog({
-    Key key,
+    Key? key,
     this.backgroundColor,
     this.elevation,
     this.insetAnimationDuration = const Duration(milliseconds: 100),
@@ -592,7 +575,7 @@ class CustomDialog extends StatelessWidget {
   ///
   /// If `null`, [ThemeData.cardColor] is used.
   /// {@endtemplate}
-  final Color backgroundColor;
+  final Color? backgroundColor;
 
   /// {@template flutter.material.dialog.elevation}
   /// The z-coordinate of this [Dialog].
@@ -601,7 +584,7 @@ class CustomDialog extends StatelessWidget {
   /// dialog's elevation is 24.0.
   /// {@endtemplate}
   /// {@macro flutter.material.material.elevation}
-  final double elevation;
+  final double? elevation;
 
   /// The duration of the animation to show when the system keyboard intrudes
   /// into the space that the dialog is placed in.
@@ -625,12 +608,12 @@ class CustomDialog extends StatelessWidget {
   ///
   /// The default shape is a [RoundedRectangleBorder] with a radius of 2.0.
   /// {@endtemplate}
-  final ShapeBorder shape;
+  final ShapeBorder? shape;
 
   /// The widget below this widget in the tree.
   ///
   /// {@macro flutter.widgets.child}
-  final Widget child;
+  final Widget? child;
 
   // TODO(johnsonmh): Update default dialog border radius to 4.0 to match material spec.
   static const RoundedRectangleBorder _defaultDialogShape =
@@ -654,7 +637,7 @@ class CustomDialog extends StatelessWidget {
         context: context,
         child: Center(
           child: ConstrainedBox(
-            constraints: BoxConstraints(minWidth: minWidth ?? 280.0),
+            constraints: BoxConstraints(minWidth: minWidth),
             child: Material(
               color: backgroundColor ??
                   dialogTheme.backgroundColor ??
